Refactor theme-graph to allow for `updateGraph`
