+++ packages/theme-graph/src/graph/build.spec.ts
@@ -10,24 +10,19 @@ import { NodeFileSystem } from '@shopify/theme-check-node';
 import path from 'node:path';
 import { assert, beforeAll, describe, expect, it } from 'vitest';
 import { URI } from 'vscode-uri';
-import { getWebComponentMap, WebComponentMap } from './getWebComponentMap';
+import { getWebComponentMap } from '../getWebComponentMap';
+import { buildThemeGraph, toSourceCode } from '../index';
 import {
-  bind,
-  buildThemeGraph,
   Dependencies,
-  getSectionModule,
-  getSnippetModule,
-  getTemplateModule,
   JsonModuleKind,
   LiquidModuleKind,
   ModuleType,
-  serializeThemeGraph,
-  ThemeGraph,
-  toSourceCode,
   Range,
-} from './index';
+  ThemeGraph,
+  WebComponentMap,
+} from '../types';
 
-const fixturesRoot = path.join(__dirname, '../fixtures');
+const fixturesRoot = path.join(__dirname, '../../fixtures');
 const skeleton = path.join(fixturesRoot, 'skeleton');
 const identity = <T>(x: T): T => x;
 
@@ -82,9 +77,17 @@ describe('Module: index', () => {
         expect(graph.rootUri).toBe(rootUri);
       });
 
-      it('infers entry points from the templates folder', () => {
-        expect(graph.entryPoints).toHaveLength(1);
-        expect(graph.entryPoints[0].uri).toBe(p('templates/index.json'));
+      // We're using sections as entry points because the section rendering API can render
+      // any section without it needing a preset or default value in its schema.
+      it('infers entry points from the templates folder and section files', () => {
+        expect(graph.entryPoints).toHaveLength(3);
+        expect(graph.entryPoints.map((x) => x.uri)).toEqual(
+          expect.arrayContaining([
+            p('templates/index.json'),
+            p('sections/custom-section.liquid'),
+            p('sections/header.liquid'),
+          ]),
+        );
       });
 
       it("finds layout/theme.liquid's dependencies and references", () => {
@@ -267,62 +270,6 @@ describe('Module: index', () => {
           ]),
         );
       });
-    });
-  });
-
-  describe('Unit: serializeThemeGraph', () => {
-    it('serialize the graph', () => {
-      const rootUri = 'file:///theme';
-      const p = (part: string) => pathUtils.join(rootUri, part);
-      const graph: ThemeGraph = {
-        entryPoints: [],
-        modules: {},
-        rootUri,
-      };
-
-      const template = getTemplateModule(graph, p('templates/index.json'));
-      const customSection = getSectionModule(graph, 'custom-section');
-      const parentSnippet = getSnippetModule(graph, 'parent');
-      const childSnippet = getSnippetModule(graph, 'child');
-      bind(template, customSection, { sourceRange: [0, 5] });
-      bind(customSection, parentSnippet, { sourceRange: [10, 15] });
-      bind(parentSnippet, childSnippet, { sourceRange: [20, 25] });
-
-      const section2 = getSectionModule(graph, 'section2');
-      bind(template, section2, { sourceRange: [20, 25] });
-
-      graph.entryPoints = [template];
-      [template, customSection, section2, parentSnippet, childSnippet].forEach((module) => {
-        graph.modules[module.uri] = module;
-      });
-
-      const { nodes, edges } = serializeThemeGraph(graph);
-      expect(nodes).toHaveLength(5);
-      expect(edges).toHaveLength(4);
-      expect(edges).toEqual(
-        expect.arrayContaining([
-          {
-            source: { uri: 'file:///theme/templates/index.json', range: [0, 5] },
-            target: { uri: 'file:///theme/sections/custom-section.liquid' },
-            indirect: false,
-          },
-          {
-            source: { uri: 'file:///theme/sections/custom-section.liquid', range: [10, 15] },
-            target: { uri: 'file:///theme/snippets/parent.liquid' },
-            indirect: false,
-          },
-          {
-            source: { uri: 'file:///theme/snippets/parent.liquid', range: [20, 25] },
-            target: { uri: 'file:///theme/snippets/child.liquid' },
-            indirect: false,
-          },
-          {
-            source: { uri: 'file:///theme/templates/index.json', range: [20, 25] },
-            target: { uri: 'file:///theme/sections/section2.liquid' },
-            indirect: false,
-          },
-        ]),
-      );
     });
   });
